requirements.txt
environs==14.1.1
protobuf
keelson==0.3.9rc15
pyserial
eclipse-zenoh==1.2.1
ouster-sdk==0.14.0
# ouster-sdk==0.11.1
